Extract Wire Protocol Snippets as method object
